refactor: move saveDraft answer and submitAnswer API from submission to answer namespace
