+++ client/app/api/course/Assessment/Submission/Answer/index.js
@@ -1,9 +1,11 @@
+import AnswersAPI from './Answer';
 import ForumPostResponseAPI from './ForumPostResponse';
 import ProgrammingAPI from './Programming';
 import ScribingsAPI from './Scribing';
 import TextResponseAPI from './TextResponse';
 
 const AnswerAPI = {
+  answer: new AnswersAPI(),
   scribing: new ScribingsAPI(),
   programming: new ProgrammingAPI(),
   textResponse: new TextResponseAPI(),
